충돌해결 commit

+++ golboogi/src/main/java/com/gf/golboogi/controller/AdminController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import com.gf.golboogi.entity.GolfFieldDto;
 import com.gf.golboogi.entity.GolfManagerDto;
 import com.gf.golboogi.entity.MemberDto;
+import com.gf.golboogi.error.CannotFindException;
 import com.gf.golboogi.repository.AdminDao;
 import com.gf.golboogi.repository.GolfFieldDao;
 import com.gf.golboogi.vo.AdminVO;
@@ -32,37 +33,57 @@ public class AdminController {
 	private GolfFieldDao golfFieldDao;
 	
 	@GetMapping("/list")
-	public String list(Model model) {
-		List<AdminVO> vo = adminDao.list();
-		model.addAttribute("adminVO", vo);
-		//
-		List<GolfFieldDto> golfFieldDto = golfFieldDao.searchSimple();
-		log.debug("DTO = {}", golfFieldDto);
-		model.addAttribute("golfFieldDto", golfFieldDto);
-		//
-		return "admin/list";
+	public String list(Model model,HttpSession session) {
+		String adminId = (String)session.getAttribute("adminLogin");
+		boolean adminCheck = adminDao.MyCheck(adminId);
+		if(adminCheck) {
+			List<AdminVO> vo = adminDao.list();
+			model.addAttribute("adminVO", vo);
+			//
+			List<GolfFieldDto> golfFieldDto = golfFieldDao.searchSimple();
+			log.debug("DTO = {}", golfFieldDto);
+			model.addAttribute("golfFieldDto", golfFieldDto);
+			//
+			return "admin/list";
+		}else {
+			throw new CannotFindException();
+		}
 	}
 	
 	@GetMapping("/detail")
-	public String detail(@RequestParam String golfManagerId, Model model ) {
-		AdminVO adminVO = adminDao.detail(golfManagerId);
-		model.addAttribute("adminVO",adminVO);
-		return "admin/detail";
+	public String detail(@RequestParam String golfManagerId, Model model,HttpSession session) {
+		String adminId = (String)session.getAttribute("adminLogin");
+		boolean adminCheck = adminDao.MyCheck(adminId);
+		if(adminCheck) {
+			AdminVO adminVO = adminDao.detail(golfManagerId);
+			model.addAttribute("adminVO",adminVO);
+			return "admin/detail";
+		}else {
+			throw new CannotFindException();
+		}
 	}
 	
 	@GetMapping("/delete")
-	public String delete(@RequestParam String golfManagerId) {
-		boolean success = adminDao.delete(golfManagerId);
-		if(success) {
+	public String delete(@RequestParam String golfManagerId,HttpSession session) {
+		String adminId = (String)session.getAttribute("adminLogin");
+		boolean adminCheck = adminDao.MyCheck(adminId);
+		if(adminCheck) {
+			adminDao.delete(golfManagerId);
 			return "redirect:list";
-		} else {
-			return "redirect:list?error";
+		}else {
+			throw new CannotFindException();
 		}
 	}
 	
 	@GetMapping("/insert")
-	public String insert() {
-		return "admin/insert";
+	public String insert(HttpSession session) {
+		String adminId = (String)session.getAttribute("adminLogin");
+		boolean adminCheck = adminDao.MyCheck(adminId);
+		if(adminCheck) {
+			return "admin/insert";
+		}else {
+			throw new CannotFindException();
+		}
 	}
 	
 //	@PostMapping("/insert")
@@ -84,24 +105,42 @@ public class AdminController {
 	}
 	
 	@GetMapping("/member_list")
-	public String memberList(Model model) {
-		List<MemberDto> memberList = adminDao.memberList();
-		model.addAttribute("memberList" , memberList);
-		return "admin/member_list";
+	public String memberList(Model model,HttpSession session) {
+		String adminId = (String)session.getAttribute("adminLogin");
+		boolean adminCheck = adminDao.MyCheck(adminId);
+		if(adminCheck) {
+			List<MemberDto> memberList = adminDao.memberList();
+			model.addAttribute("memberList" , memberList);
+			return "admin/member_list";
+		}else {
+			throw new CannotFindException();
+		}
 	}
 	
 	@GetMapping("/member_detail")
-	public String memberDetail(@RequestParam String memberId, Model model) {
-		MemberDto memberDto = adminDao.memberDetail(memberId);
-		model.addAttribute("memberDto", memberDto);
-		return "admin/member_detail";
+	public String memberDetail(@RequestParam String memberId, Model model,HttpSession session) {
+		String adminId = (String)session.getAttribute("adminLogin");
+		boolean adminCheck = adminDao.MyCheck(adminId);
+		if(adminCheck) {
+			MemberDto memberDto = adminDao.memberDetail(memberId);
+			model.addAttribute("memberDto", memberDto);
+			return "admin/member_detail";
+		}else {
+			throw new CannotFindException();
+		}
 	}
 	
 	@GetMapping("/member_blacklist")
-	public String memberBlacklist(@RequestParam String memberId, Model model) {
-		MemberDto memberDto = adminDao.memberDetail(memberId);
-		model.addAttribute("memberDto", memberDto);
-		return "admin/member_blacklist";
+	public String memberBlacklist(@RequestParam String memberId, Model model,HttpSession session) {
+		String adminId = (String)session.getAttribute("adminLogin");
+		boolean adminCheck = adminDao.MyCheck(adminId);
+		if(adminCheck) {
+			MemberDto memberDto = adminDao.memberDetail(memberId);
+			model.addAttribute("memberDto", memberDto);
+			return "admin/member_blacklist";
+		}else {
+			throw new CannotFindException();
+		}
 	}
 	
 	@PostMapping("/member_blacklist")
@@ -138,9 +177,9 @@ public class AdminController {
 	
 	@RequestMapping("/logout")
 	public String logout(HttpSession session) {
-		session.removeAttribute("adminLogin");
-		session.removeAttribute("auth");
-		return "redirect:/";
+			session.removeAttribute("adminLogin");
+			session.removeAttribute("auth");
+			return "redirect:/";
 	}
