+++ golboogi/src/main/java/com/gf/golboogi/controller/ReviewController.java
@@ -45,7 +45,6 @@ public class ReviewController {
 	@Autowired
 	private AttachmentDao attachmentDao;
 	
-	
 	@GetMapping("/list")
 	public String list(
 			@RequestParam(required = false) String keyword,
@@ -141,11 +140,18 @@ public class ReviewController {
 	}
 	
 	@GetMapping("/edit/{reviewNo}")
-	public String edit(@PathVariable int reviewNo, Model model) {
-		ReviewDto reviewDto = reviewDao.info(reviewNo);
-		model.addAttribute("reviewDto", reviewDto);
+	public String edit(@PathVariable int reviewNo, Model model,HttpSession session) {
+		String memberId = (String)session.getAttribute("login");
+		String reviewWriter = reviewDao.writerCheck(memberId);
+		boolean MyCheck = reviewWriter == memberId;
+		if(MyCheck) {
+			ReviewDto reviewDto = reviewDao.info(reviewNo);
+			model.addAttribute("reviewDto", reviewDto);
+			return "review/edit";
+		}else {
+			throw new CannotFindException();
+		}
 		
-		return "review/edit";
 	}
 	
 	@PostMapping("/edit/{reviewNo}")
@@ -194,12 +200,14 @@ public class ReviewController {
 	}
 	
 	@GetMapping("/delete/{reviewNo}")
-	public String delete(@PathVariable int reviewNo) {
-		boolean success = reviewDao.delete(reviewNo);
-		if(success) {
+	public String delete(@PathVariable int reviewNo,HttpSession session) {
+		String memberId = (String)session.getAttribute("login");
+		String reviewWriter = reviewDao.writerCheck(memberId);
+		boolean deleteCheck = memberId == reviewWriter;
+		if(deleteCheck) {
+			reviewDao.delete(reviewNo);
 			return "redirect:/review/list";
-		}
-		else {
+		}else {
 			throw new CannotFindException();
 		}
 	}
